pyproject.toml
[project]
name = "career-twin"
version = "0.1.0"
description = "AI-powered chatbot representing Samuel Shulman to recruiters"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.0.0",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=0.2.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.2.1",
]
